Removed tracing middleware which should be in graphql-server-core
